security/openvpn: allow build with libressl+LZO
Just drop the code enforcing the restriction for now. This is NOT ready
for submission.

GPL allows the user to do "whatever" with the code as long as they don't
publish a changed version or derived work, which is when restrictions
apply. So, I'm pretty sure this includes linking whatever libs you want.

There *should* be a restriction, but it should probably only forbid to
build official packages, not to build locally for your own use.