lang/dotnet: Add experimental LibreSSL patch
Upstream source does not support LibreSSL. It can be patched to build
with LibreSSL, but this requires disabling some functions and also
interfers with the runtime OpenSSL version detection.

Therefore add a port option LIBRESSL for this, marked as experimental
and explained in pkg-descr. Mark the port as broken with
SSL_DEFAULT=libressl when this option is not set.